requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.23.1
anyio==4.8.0
argcomplete==3.5.3
astor==0.8.1
attrs==25.1.0
backoff==2.2.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-option-group==0.5.6
codeowners==0.7.0
cryptography==44.0.1
dataclasses-json==0.6.7
datamodel-code-generator==0.28.1
Deprecated==1.2.18
dicttoxml==1.7.16
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
dotty-dict==1.3.1
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.44
giturlparse==0.12.0
greenlet==3.1.1
grpclib==0.4.7
h11==0.14.0
h2==4.2.0
hatch-vcs==0.4.0
hatchling==1.27.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanize==4.12.0
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
inflect==5.6.2
iniconfig==2.0.0
isort==6.0.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain
langchain-openai
langgraph
openai
lazy-object-proxy==1.10.0
Levenshtein==0.26.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.2.1
mdurl==0.1.2
mini-racer==0.12.4
modal==0.73.51
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.26.0
neo4j==5.28.1
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.3
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pygit2==1.17.0
PyGithub==2.5.0
Pygments==2.19.1
pyinstrument==5.0.1
pyjson5==1.6.8
PyJWT==2.10.1
PyNaCl==1.5.0
pyright==1.1.394
pytest==8.3.4
pytest-snapshot==0.9.0
python-dotenv
python-gitlab==4.13.0
python-Levenshtein==0.26.1
python-multipart==0.0.20
python-semantic-release==9.19.1
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-click==1.8.5
rich-toolkit==0.13.2
rustworkx==0.16.0
sentry-sdk==2.20.0
setuptools==75.8.0
setuptools-scm==8.1.0
shellingham==1.5.4
sigtools==4.0.1
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sse-starlette==2.2.1
starlette==0.45.3
synchronicity==0.9.11
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
tree-sitter==0.24.0
tree-sitter-javascript==0.23.1
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2
trove-classifiers==2025.1.15.22
typer==0.15.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspect==0.9.0
typing_extensions==4.12.2
unidiff==0.7.5
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
wrapt==1.17.2
xmltodict==0.14.2
yarl==1.18.3
zstandard==0.23.0
python-pptx
pyppeteer
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
matplotlib
graphviz
gdown